requirements.txt
numpy=1.21.5
ffmpeg-python
playsound==1.2.2
pyaudio
colorama
